fix(types): fix all remaining basedpyright type errors
Fixed 102 type errors across multiple files:

- video_writer.py: PyAV dynamic attributes (type: ignore[attr-defined])
- gradio_helpers.py: Path vs str type annotations, MatLike arg-type ignores
- phase_detector.py: numpy scalar→python int/float casts
- pose_tracker.py: None guards for optional ndarray attributes
- detection/person_detector.py: type ignore for dynamic import
- visualization layers: arg-type ignores for Position2D/ndarray
- tracking/: None guards for optional attributes
- pose_3d/: None guards for model, import ignores, dtype fixes
- utils/geometry.py: .astype(float32) for return type consistency
- detection/spatial_reference.py: float() cast, None guard

All files now pass basedpyright --level error with 0 errors.

+++ src/utils/geometry.py
@@ -129,7 +129,7 @@ def get_mid_hip(poses: NormalizedPose) -> NDArray[np.float32]:
     Returns:
         Mid-hip coordinates (num_frames, 2).
     """
-    return (poses[:, H36Key.LHIP, :] + poses[:, H36Key.RHIP, :]) / 2
+    return ((poses[:, H36Key.LHIP, :] + poses[:, H36Key.RHIP, :]) / 2).astype(np.float32)
 
 
 def get_mid_shoulder(poses: NormalizedPose) -> NDArray[np.float32]:
@@ -141,7 +141,7 @@ def get_mid_shoulder(poses: NormalizedPose) -> NDArray[np.float32]:
     Returns:
         Mid-shoulder coordinates (num_frames, 2).
     """
-    return (poses[:, H36Key.LSHOULDER, :] + poses[:, H36Key.RSHOULDER, :]) / 2
+    return ((poses[:, H36Key.LSHOULDER, :] + poses[:, H36Key.RSHOULDER, :]) / 2).astype(np.float32)
 
 
 def calculate_center_of_mass(poses: NormalizedPose, frame_idx: int) -> float:
